Explicit conversion to int

+++ include/nlohmann/detail/output/serializer.hpp
@@ -160,7 +160,7 @@ class serializer
                     return;
                 }
 
-                if (pretty_print(val, current_indent))
+                if (pretty_print(val, static_cast<int>(current_indent)))
                 {
                     o->write_characters("{\n", 2);
 
@@ -233,7 +233,7 @@ class serializer
                     return;
                 }
 
-                if (pretty_print(val, current_indent))
+                if (pretty_print(val, static_cast<int>(current_indent)))
                 {
                     o->write_characters("[\n", 2);
 
@@ -295,7 +295,7 @@ class serializer
 
             case value_t::binary:
             {
-                if (pretty_print(val, current_indent))
+                if (pretty_print(val, static_cast<int>(current_indent)))
                 {
                     o->write_characters("{\n", 2);
